requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2023.11.16
spotdl==4.2.4
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.4.2
python-dotenv==1.0.0
